extract RecordingMocks helper class
