+++ src/test/java/org/cru/redegg/RedEggAppenderIntegrationTest.java
@@ -2,8 +2,6 @@ package org.cru.redegg;
 
 
 import org.cru.redegg.boot.Lifecycle;
-import org.cru.redegg.recording.api.NoOpParameterSanitizer;
-import org.cru.redegg.recording.api.ParameterSanitizer;
 import org.cru.redegg.recording.api.RecorderFactory;
 import org.cru.redegg.recording.api.WebErrorRecorder;
 import org.cru.redegg.recording.jul.RedEggHandler;
@@ -22,21 +20,13 @@ import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Mock;
-import org.mockito.Mockito;
-import org.mockito.MockitoAnnotations;
 
-import javax.annotation.PostConstruct;
-import javax.enterprise.context.ApplicationScoped;
-import javax.enterprise.inject.Produces;
 import javax.inject.Inject;
 import java.util.logging.LogRecord;
 
 import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.atLeast;
-import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 
 @RunWith(Arquillian.class)
 public class RedEggAppenderIntegrationTest
@@ -59,7 +49,8 @@ public class RedEggAppenderIntegrationTest
             .addClass(LoggingReporter.class)
             .addClass(ErrorReporter.class)
 
-            .addClass(AnswerWithSelf.class);
+            .addClass(AnswerWithSelf.class)
+            .addClass(RecordingMocks.class);
     }
 
 
@@ -67,7 +58,7 @@ public class RedEggAppenderIntegrationTest
     WebErrorRecorder recorder;
 
 
-    @Inject Mocks mocks;
+    @Inject RecordingMocks mocks;
 
     @Before
     public void setup()
@@ -91,39 +82,5 @@ public class RedEggAppenderIntegrationTest
         verify(recorder, atLeast(1)).sendReportIfNecessary();
     }
 
-
-    @ApplicationScoped
-    public static class Mocks
-    {
-
-        @Produces
-        @Mock
-        RecorderFactory factory;
-
-        @Produces
-        WebErrorRecorder recorder;
-
-
-        @Produces
-        ParameterSanitizer sanitizer = new NoOpParameterSanitizer();
-
-        @PostConstruct
-        public void init()
-        {
-            MockitoAnnotations.initMocks(this);
-            recorder = mock(WebErrorRecorder.class, new AnswerWithSelf(WebErrorRecorder.class));
-            reset();
-        }
-
-        public void reset()
-        {
-            // we use Mockito.reset() instead of building new mocks, because the servlet listener is only initialized
-            // once for this test class, and there is no easy way to modify its reference to a new mock
-            Mockito.reset(recorder, factory);
-            when(factory.getRecorder()).thenReturn(recorder);
-            when(factory.getWebRecorder()).thenReturn(recorder);
-        }
-
-    }
 
 }
